test/wpt/wpt-alt/link-pseudo-in-has.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>CSS Selectors Invalidation: :link, :visited :any-link, pseudo-class in :has() argument</title>
<link rel="author" title="Byungwoo Lee" href="blee@igalia.com">
<link rel="help" href="https://drafts.csswg.org/selectors/#relational">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #parent { color: blue; }
  #grandparent { color: blue; }
  #parent:has(> :not(:link)) { color: grey; }
  #parent:has(> :link) { color: green; }
  #parent:has(> :visited) { color: red; }
  #grandparent:has(:not(:any-link)) { color: grey; }
  #grandparent:has(:any-link) { color: green; }
</style>
<div id="grandparent"></div>
<script>
  const BLUE = "rgb(0, 0, 255)";
  const GREY = "rgb(128, 128, 128)";
  const GREEN = "rgb(0, 128, 0)";
  const RED = "rgb(255, 0, 0)";

  function checkColor(id, color, target_matches) {
    let element = document.getElementById(id);
    let message = ["location.hash ==", location.hash, ": #" + id, "should be",
                   color, (target_matches ? "with" : "without"),
                   ":target"].join(" ");
    assert_equals(getComputedStyle(element).color, color, message);
  }

  promise_test(async () => {
    assert_equals(grandparent.matches('#grandparent:has(:not(:any-link))'),
      false, "grandparent should be blue without any element");
    assert_equals(grandparent.matches('#grandparent:has(:any-link)'),
      false, "grandparent should be blue without any element 2");

    let parent = document.createElement("div");
    parent.id = "parent";
    grandparent.appendChild(parent);

    assert_equals(grandparent.matches('#grandparent:has(:not(:any-link))'),
      true, "grandparent should be grey after parent added");
    assert_equals(grandparent.matches('#grandparent:has(:any-link)'),
      false, "grandparent should be grey after parent added 2");
    assert_equals(parent.matches('#parent:has(> :not(:link))'), false,
      "parent should be blue without any link");
    assert_equals(parent.matches('#parent:has(> :link)'), false,
      "parent should be blue without any link 2");
    assert_equals(parent.matches('#parent:has(> :visited)'), false,
      "parent should be blue without any link 3");

    let div = document.createElement("div");
    parent.appendChild(div);

    assert_equals(grandparent.matches('#grandparent:has(:not(:any-link))'),
      true, "grandparent should be grey after div added");
    assert_equals(grandparent.matches('#grandparent:has(:any-link)'),
      false, "grandparent should be grey after div added 2");
    assert_equals(parent.matches('#parent:has(> :not(:link))'), true,
      "parent should be grey after div added");
    assert_equals(parent.matches('#parent:has(> :link)'), false,
      "parent should be grey after div added 2");
    assert_equals(parent.matches('#parent:has(> :visited)'), false,
      "parent should be grey after div added 3");

    let visited = document.createElement("a");
    visited.href = "";
    parent.appendChild(visited);

    assert_equals(grandparent.matches('#grandparent:has(:any-link)'),
      true, "grandparent should be green after visited link added");
    assert_equals(parent.matches('#parent:has(> :link)'), true,
      "parent should be green after visited link added");
    assert_equals(parent.matches('#parent:has(> :visited)'), false,
      "parent should be green after visited link added 2");

    let unvisited = document.createElement("a");
    unvisited.href = "unvisited";
    parent.appendChild(unvisited);

    assert_equals(grandparent.matches('#grandparent:has(:any-link)'), true,
      "grandparent should be green after unvisited link added");
    assert_equals(parent.matches('#parent:has(> :link)'), true,
      "parent should be green after unvisited link added");
    assert_equals(parent.matches('#parent:has(> :visited)'), false,
      "parent should be green after unvisited link added 2");

    unvisited.remove();

    assert_equals(grandparent.matches('#grandparent:has(:any-link)'), true,
      "grandparent should be green after unvisited link removed");
    // FIXME: why `false` even though a child div exists in #parent?
    /*
    assert_equals(parent.matches('#parent:has(> :not(:link))'), false,
      "parent should be blue after unvisited link removed");
    // NOTE: fails
    assert_equals(parent.matches('#parent:has(> :link)'), false,
      "parent should be blue after unvisited link removed 2");
    */
    assert_equals(parent.matches('#parent:has(> :visited)'), false,
      "parent should be blue after unvisited link removed 3");

    visited.remove();

    assert_equals(grandparent.matches('#grandparent:has(:not(:any-link))'),
      true, "grandparent should be grey after visited link removed");
    assert_equals(grandparent.matches('#grandparent:has(:any-link)'),
      false, "grandparent should be grey after visited link removed 2");
    assert_equals(parent.matches('#parent:has(> :not(:link))'), true,
      "parent should be grey after visited link removed");
    assert_equals(parent.matches('#parent:has(> :link)'), false,
      "parent should be grey after visited link removed 2");
    assert_equals(parent.matches('#parent:has(> :visited)'), false,
      "parent should be grey after visited link removed 3");

    div.remove();

    assert_equals(grandparent.matches('#grandparent:has(:not(:any-link))'),
      true, "grandparent should be grey after div removed");
    assert_equals(grandparent.matches('#grandparent:has(:any-link)'),
      false, "grandparent should be grey after div removed 2");
    assert_equals(parent.matches('#parent:has(> :not(:link))'), false,
      "parent should be blue after div removed");
    assert_equals(parent.matches('#parent:has(> :link)'), false,
      "parent should be blue after div removed 2");
    assert_equals(parent.matches('#parent:has(> :visited)'), false,
      "parent should be blue after div removed 3");

    parent.remove();

    assert_equals(grandparent.matches('#grandparent:has(:not(:any-link))'),
      false, "grandparent should be blue after parent removed");
    assert_equals(grandparent.matches('#grandparent:has(:any-link)'),
      false, "grandparent should be blue after parent removed 2");
  });
</script>
